w3c-reports/12092014/efl/runned-tests/IndexedDB/idbcursor-direction-index-expected.txt

CONSOLE MESSAGE: line 24: ReferenceError: Can't find variable: indexedDB

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: indexedDB
